test: 빌드 테스트 보완
빌드 테스트 보완
  - 보완점 정리
    - success 리다이렉트 지연 제거

    - 회원가입 시 토근 제거, 회원가입 -> 로그인 흐름 개선

    - 메뉴 스크롤 범위 확장, 메인 섹션 너비 조건부 할당

    - 메뉴 카테고리 오류 알림 명시, 토스트 알림 -> 글자 표시 적용

    - 좌석 편집 배경 퇴장 시기 조정, 확인 모달 제출 후 퇴장 적용

    - 좌석 요청 알림 문구 오류 수정

권한 관리
  - 사용자 목록 관리, auth 연동 테이블 생성

  - 익명 사용자 로그인 주소 반환, raw_user_meta_data 설정

+++ src/features/modal/tab/menu/update/index.tsx
@@ -72,7 +72,7 @@ export default function UpdateMenuModal() {
           await updateImage({ file: menuImageFile, fileId });
         }
 
-        if (submitType === 'delete') {
+        if (submitType === 'delete' && fileId !== 'default') {
           const filePath = [STORE + `/menu_${fileId}`];
           await deleteImageByFileName({ filePath });
         }
